pull out a few basic functions in their own modules
